Enabled adding, deleting and editing Header Content. Created global so that keys for the json file could be identified by the IDs. Restructured how Adding content works. Enabled additions of code sections. Added a very simple img content adding feature.
Unable to add a header section as content. Getting the keys was becoming complex and looked awful. Adding content was
done with individual functions for each type of content, which was messy and repeated code.

Added in functionality to the Add header selection on the context menu for the page content. Created a global object
to store the id/ key values as getting keys for the json file was getting messy. Created an adding content function
to handle adding all types of content.